Move ttrpc client to pkg/ttrpcutil
