+++ pkg/ttrpcutil/client_unix.go
@@ -16,7 +16,7 @@
    limitations under the License.
 */
 
-package containerd
+package ttrpcutil
 
 import (
 	"net"
