+++ pkg/ttrpcutil/client_windows.go
@@ -16,7 +16,7 @@
    limitations under the License.
 */
 
-package containerd
+package ttrpcutil
 
 import (
 	"net"
